digb_backend/requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.24.0.post1  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==4.2.7  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.58.2  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2023.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.3.1  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.5  # https://github.com/tfranzel/drf-spectacular

#dependencies added from pip freeze

anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==2023.10
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-json-widget==1.1.1
django-model-utils==4.3.1
django-redis==5.4.0
django-sites==0.11
django-utils-six==2.0
djangorestframework==3.14.0
drf-core==0.0.4
drf-spectacular==0.26.5
exceptiongroup==1.2.0
future==0.18.3
h11==0.14.0
hiredis==2.2.3
httptools==0.6.1
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
oauthlib==3.2.2
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0
zipp==3.17.0
